Rename `add-site-blueprint-from-url` to `add-site-with-blueprint`
