+++ src/ipc-utils.ts
@@ -19,7 +19,7 @@ type SnapshotKeyValueEventData = {
 
 export interface IpcEvents {
 	'add-site': [ void ];
-	'add-site-blueprint-from-url': [ { blueprintPath: string } ];
+	'add-site-with-blueprint': [ { blueprintPath: string } ];
 	'auth-updated': [ { token: StoredToken } | { error: unknown } ];
 	'on-export': [ ImportExportEventData, string ];
 	'on-import': [ ImportExportEventData, string ];
